graph/requirements.txt
streamlit
# 사용할 패키지/라이브러리
numpy
pandas
seaborn
matplotlib.pyplot
